"feat: implemented ThemeServices, refactored code"
